+++ jnpr/openclos/underlayRestRoutes.py
@@ -43,6 +43,7 @@ class UnderlayRestRoutes():
         self.__dao = context['dao']
         self.__daoClass = context['daoClass']
         self.app = context['app']
+        self.restServer = context['restServer']
 
         if 'outputDir' in self._conf:
             global webServerRoot
@@ -88,7 +89,9 @@ class UnderlayRestRoutes():
         self.app.route(self.baseUrl + '/pods/<podId>', 'DELETE', self.deletePod)
 
     def getPods(self, dbSession):
-        
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         url = str(bottle.request.url).translate(None, ',')
         podsData = {}
         listOfIpFbarics = []
@@ -120,6 +123,9 @@ class UnderlayRestRoutes():
                 'outOfBandAddressList', 'outOfBandGateway', 'topologyType', 'spineJunosImage', 'hostOrVmCountPerLeaf']
     
     def getPod(self, dbSession, podId, requestUrl=None):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         if requestUrl is None:
             requestUrl = str(bottle.request.url).translate(None, ',')
         pod = self.report.getPod(dbSession, podId)
@@ -169,7 +175,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(404, exception=PodNotFound(podId))
     
     def getCablingPlan(self, dbSession, podId):
-        
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         header = bottle.request.get_header('Accept')
         logger.debug('Accept header before processing: %s', header)
         # hack to remove comma character, must be a bug on Bottle
@@ -203,6 +211,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(404, exception=PodNotFound(podId))
 
     def getLeafGenericConfiguration(self, dbSession, podId, deviceModel):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         pod = self.report.getPod(dbSession, podId)
         if pod is None:
             raise bottle.HTTPError(404, exception=PodNotFound(podId))
@@ -218,6 +229,9 @@ class UnderlayRestRoutes():
         return leafSetting.config
 
     def getDeviceConfigsInZip(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         pod = self.report.getPod(dbSession, podId)
         if pod is None:
             raise bottle.HTTPError(404, exception=PodNotFound(podId))
@@ -254,7 +268,9 @@ class UnderlayRestRoutes():
         Hook to enhance Device object
         '''
     def getDevices(self, dbSession, podId):
-        
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         devices = {}
         listOfDevices = []
         pod = self.report.getPod(dbSession, podId)
@@ -284,6 +300,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(404, exception=PodNotFound(podId))
         
     def getDevice(self, dbSession, podId, deviceId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         device = UnderlayRestRoutes.isDeviceExists(dbSession, podId, deviceId)
         #podUri is constructed from url
         url = str(bottle.request.url).translate(None, ',')
@@ -322,7 +341,9 @@ class UnderlayRestRoutes():
         
          
     def getDeviceConfig(self, dbSession, podId, deviceId):
-        
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         device = UnderlayRestRoutes.isDeviceExists(dbSession, podId, deviceId)
         if device is None:
             raise bottle.HTTPError(404, exception=DeviceNotFound("No device found with podId: '%s', deviceId: '%s'" % (podId, deviceId)))
@@ -336,7 +357,9 @@ class UnderlayRestRoutes():
 
     
     def getZtpConfig(self, dbSession, podId):
-        
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         pod = self.report.getPod(dbSession, podId)
         if pod is not None:
             logger.debug('pod name: %s', pod.name)
@@ -360,7 +383,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(404, exception=DeviceNotFound("No device found with podId: '%s', deviceId: '%s'" % (podId, deviceId)))
 
     def getJunosImage(self, dbSession, junosImageName):
-           
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         fileName = os.path.join(junosImageRoot, junosImageName)
         logger.debug('junosImageRoot: %s, image: %s, exists: %s', junosImageRoot, junosImageName, os.path.exists(fileName))
 
@@ -370,6 +395,9 @@ class UnderlayRestRoutes():
         return config
     
     def getOpenClosConfigParams(self, dbSession):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         supportedDevices = []
         
         for deviceFamily, value in self.deviceSku.skuDetail.iteritems():
@@ -392,6 +420,9 @@ class UnderlayRestRoutes():
         return {'OpenClosConf' : confValues}
                     
     def createPod(self, dbSession):  
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         if bottle.request.json is None:
             raise bottle.HTTPError(400, exception=InvalidRequest("No json in request object"))
         else:
@@ -416,6 +447,9 @@ class UnderlayRestRoutes():
         return pod
         
     def createCablingPlan(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         try:
             l3ClosMediation = L3ClosMediation(self._conf, self.__daoClass)
             if l3ClosMediation.createCablingPlan(podId) is True:
@@ -426,6 +460,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(500, exception=exc)
 
     def createDeviceConfiguration(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         try:
             l3ClosMediation = L3ClosMediation(self._conf, self.__daoClass)
             if l3ClosMediation.createDeviceConfig(podId) is True:
@@ -436,6 +473,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(500, exception=exc)
             
     def createZtpConfiguration(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         try:
             ZtpServer().createPodSpecificDhcpConfFile(dbSession, podId)
         except PodNotFound as exc:
@@ -444,6 +484,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(500, exception=exc)
 
     def reconfigPod(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         if bottle.request.json is None:
             raise bottle.HTTPError(400, exception=InvalidRequest("No json in request object"))
         else:
@@ -465,9 +508,15 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(400, exception=exc)
     
     def setOpenClosConfigParams(self):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         return bottle.HTTPResponse(status=200)
     
     def deletePod(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         pod = self.report.getPod(dbSession, podId)
         if pod is not None:
             self.__dao.deleteObject(dbSession, pod)
@@ -548,6 +597,9 @@ class UnderlayRestRoutes():
         return podDevices
 
     def getL2Report(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         try:
             cached = bottle.request.query.get('cached', '1')
             if cached == '1':
@@ -561,6 +613,9 @@ class UnderlayRestRoutes():
             raise bottle.HTTPError(404, exception=PodNotFound(podId, exc))
     
     def getL3Report(self, dbSession, podId):
+        if not self.restServer.checkPass():
+            raise bottle.HTTPError(401)
+            
         try:
             cached = bottle.request.query.get('cached', '1')
             if cached == '1':
